Patchtester icon
Currently we have no icon at the toolbar. This PR will change (or more fallback) to a circle with a point in it. As the others that are in the images folder don't look so good.

But if we want to have e.g. the red beetle (https://github.com/joomla-extensions/patchtester/tree/master/administrator/components/com_patchtester/assets/images)

I can change this PR

Let me know :)